chandra/ocr_service.py
# pip install pillow
from PIL import Image
from chandra.model_loader import getChandraModel
import torch

def analyzeImageWithQuestion(imagePath, userQuestion):
    """
    이미지에서 텍스트를 추출하고 사용자의 질문에 답변을 생성합니다.
    """
    try:
        model, tokenizer = getChandraModel()
        inputImage = Image.open(imagePath).convert("RGB")
        
        # Chandra OCR 모델의 특성에 따른 분석 로직 (예시 인터페이스 기반)
        # 실제 모델의 forward/generate API에 맞춰 구현이 필요합니다.
        # 여기서는 모델이 이미지와 텍스트 입력을 동시에 받는다고 가정합니다.
        
        # 텍스트 추출 및 질문 처리 예시 (가상 로직)
        # result = model.generate(inputImage, userQuestion)
        
        extractedText = "이미지에서 추출된 가상의 텍스트 내용입니다."
        generatedAnswer = f"질문 '{userQuestion}'에 대한 답변: 추출된 텍스트를 바탕으로 분석한 결과입니다."
        
        return {
            "success": True,
            "extractedText": extractedText,
            "answer": generatedAnswer
        }
    except Exception as e:
        return {
            "success": False, 
            "message": str(e)
        }

def performOcr(imagePath):
    """ 단순히 이미지의 모든 텍스트를 추출하는 함수입니다. """
    # 구현 방식은 위와 유사하며 반복문을 사용할 때 규칙을 준수합니다.
    pass
