chandra/requirements.txt
torch
transformers
Pillow
mysql-connector-python
